Fix double roles being sometimes set
Sometimes two different roles could be set due to the delay between the bot and the discord API. This could create a case where add would add back the "removed" role that wasn't actually removed in time.

Instead of using remove/add, manage the roles in-memory and set them once to the user.